requirements.txt
numpy==1.26.4
scipy==1.12.0
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
transformers==4.37.2
ultralytics==8.1.2
opencv-python==4.9.0.80
pillow==10.2.0
flask==3.0.2
redis==5.0.1
requests==2.31.0 